# docs/openapi.yaml
openapi: 3.0.3
info:
  title: TravlrAPI MVP
  description: |
    A comprehensive travel planning API that uses AI agents to generate personalized trip recommendations.
    
    The API orchestrates multiple specialized agents (flight, accommodation, activity, restaurant, transportation) 
    to create customized travel itineraries based on user preferences and budget constraints.
    
    ## Key Features
    - **Agent-based Planning**: AI agents work in parallel to find optimal recommendations
    - **Real-time Status**: Track planning progress with live agent execution updates
    - **Flexible Selection**: Users can select multiple recommendations per category
    - **Collaborative Planning**: Support for multiple users on shared trips
    - **Performance Optimized**: Sub-30 second trip planning with async execution
    
    ## Base URL
    - Development: `http://localhost:3000`
    - Production: TBD
    
    ## Authentication
    Currently anonymous access. Future versions will implement JWT authentication.
    
    ## Rate Limits
    - 100 requests per minute per IP
    - 10 trip creations per hour per IP
    - Status polling: 1 request per second
  version: 1.0.0
  contact:
    name: TravlrAPI Support
    email: support@travlr.com
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:3000
    description: Development server
  - url: https://api.travlr.com
    description: Production server (TBD)
paths:
...
  /api/trip/create:
    post:
      summary: Create Trip
      description: |
        Create a new trip and optionally trigger AI agent execution for recommendation generation.
        
        **Agent Execution Process:**
        1. Trip created with status 'planning'
        2. 5 AI agents execute in dependency-based phases:
           - Phase 1 (Parallel): Flight, Accommodation
           - Phase 2 (Sequential): Activity, Restaurant, Transportation
        3. Each agent generates 3-5 recommendations with confidence scores
        4. Trip status changes to 'recommendations_ready' when complete
        
        **Execution Time:** Typically 15-30 seconds for full agent execution.
      operationId: createTrip
      tags:
        - Trips
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateTripRequest'
            examples:
              paris_trip:
                summary: Paris vacation example
                value:
                  title: "Romantic Paris Getaway"
                  destination: "Paris"
                  origin: "New York"
                  departureDate: "2025-12-15"
                  returnDate: "2025-12-20"
                  travelers:
                    count: 2
                    adults: 2
                    children: 0
                    infants: 0
                  preferences:
                    interests: ["cultural", "food", "art"]
                    budget:
                      total: 3000
                      currency: "USD"
                      breakdown:
                        flight: 1200
                        accommodation: 800
                        food: 600
                        activities: 400
                    accommodation:
                      type: "hotel"
                      minRating: 4
                      requiredAmenities: ["wifi", "restaurant"]
                    transportation:
                      flightClass: "economy"
                      preferNonStop: true
                    dining:
                      dietaryRestrictions: ["vegetarian"]
                      priceRange: "mid_range"
                  collaboration:
                    createdBy: "user123"
              business_trip:
                summary: Business trip example
                value:
                  title: "Tokyo Business Trip"
                  destination: "Tokyo"
                  origin: "San Francisco"
                  departureDate: "2025-11-10"
                  returnDate: "2025-11-15"
                  travelers: 1
                  preferences:
                    interests: ["business", "food"]
                    budget:
                      total: 4000
                      currency: "USD"
                    accommodation:
                      type: "hotel"
                      minRating: 5
                      requiredAmenities: ["business_center", "wifi"]
                    transportation:
                      flightClass: "business"
                      preferNonStop: true
                  collaboration:
                    createdBy: "executive456"
      responses:
        '201':
          description: Trip created successfully, agent execution initiated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CreateTripResponse'
        '400':
          description: Validation errors in request body
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                success: false
                error: "Validation failed"
                details:
                  - "departureDate cannot be in the past"
                  - "travelers.count must be at least 1"
                  - "interests must contain valid enum values"
                message: "Please check your input and try again"
        '500':
          description: Internal server error during trip creation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/trip/{tripId}:
    get:
      summary: Get Trip Details
      description: |
        Retrieve complete trip information including all generated recommendations and user selections.
        
        **Populated Data:**
        - All recommendation objects with full details (not just IDs)
        - Agent execution metadata and confidence scores
        - Selected recommendations with selection timestamps
        - Trip status and collaboration information
        
        **Performance Note:** This endpoint performs multiple database joins. For frequent polling, use `/status` endpoint instead.
      operationId: getTripById
      tags:
        - Trips
      parameters:
        - name: tripId
          in: path
          required: true
          description: Unique trip identifier (generated during trip creation)
          schema:
            type: string
            pattern: '^trip_\d+_[a-z0-9]+$'
          example: "trip_1758349943478_2fq39bbwr"
      responses:
        '200':
          description: Trip details retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TripDetailsResponse'
        '404':
          description: Trip not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                success: false
                error: "Trip not found"
                message: "Trip with ID trip_invalid_id does not exist"
        '500':
          description: Database error retrieving trip
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/trip/{tripId}/status:
    get:
      summary: Get Trip Status
      description: |
        Get real-time execution status and agent progress without full recommendation data.
        
        **Optimized for Polling:**
        - Lightweight response (no recommendation details)
        - Real-time agent execution progress
        - Estimated completion times
        - Recommendation counts per category
        
        **Recommended polling interval:** 2-5 seconds during active execution.
      operationId: getTripStatus
      tags:
        - Trips
      parameters:
        - name: tripId
          in: path
          required: true
          description: Unique trip identifier
          schema:
            type: string
            pattern: '^trip_\d+_[a-z0-9]+$'
          example: "trip_1758349943478_2fq39bbwr"
      responses:
        '200':
          description: Trip status retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TripStatusResponse'
        '404':
          description: Trip not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Error retrieving trip status
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/trip/{tripId}/select:
    put:
      summary: Select Recommendations
      description: |
        Update user's recommendation selections across all categories.
        
        **Selection Rules:**
        - Can select multiple recommendations per category
        - Previously selected items are replaced (not appended)
        - Recommendation IDs must belong to the specified trip
        - Trip status must be 'recommendations_ready' or 'user_selecting'
        
        **Selection Metadata:**
        - Timestamp of selection
        - User who made the selection
        - Selection rank/order for multiple items
      operationId: selectRecommendations
      tags:
        - Trips
      parameters:
        - name: tripId
          in: path
          required: true
          description: Unique trip identifier
          schema:
            type: string
            pattern: '^trip_\d+_[a-z0-9]+$'
          example: "trip_1758349943478_2fq39bbwr"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SelectRecommendationsRequest'
            examples:
              partial_selection:
                summary: Selecting some recommendations
                value:
                  selections:
                    flight: ["67144e5c4d2b5f00128a1234"]
                    accommodation: ["67144e5c4d2b5f00128a5678", "67144e5c4d2b5f00128a9012"]
                    activity: ["67144e5c4d2b5f00128a3456"]
                  selectedBy: "user123"
              full_selection:
                summary: Complete trip selections
                value:
                  selections:
                    flight: ["67144e5c4d2b5f00128a1111"]
                    accommodation: ["67144e5c4d2b5f00128a2222"]
                    activity: ["67144e5c4d2b5f00128a3333", "67144e5c4d2b5f00128a4444"]
                    restaurant: ["67144e5c4d2b5f00128a5555"]
                    transportation: ["67144e5c4d2b5f00128a6666"]
                  selectedBy: "user456"
      responses:
        '200':
          description: Selections updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SelectRecommendationsResponse'
        '400':
          description: Invalid selections or trip not ready
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              examples:
                invalid_recommendations:
                  summary: Invalid recommendation IDs
                  value:
                    success: false
                    error: "Invalid recommendation IDs for flight"
                    details: ["67144e5c4d2b5f00128a9999"]
                    message: "Some selected recommendations do not belong to this trip"
                trip_not_ready:
                  summary: Trip not ready for selections
                  value:
                    success: false
                    error: "Trip not ready for selections"
                    message: "Please wait for recommendations to be generated before making selections"
        '404':
          description: Trip not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Error updating selections
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/trip/{tripId}/rerun:
    post:
      summary: Rerun Agents
      description: |
        Trigger selective or complete re-execution of AI agents for updated recommendations.
        
        **Use Cases:**
        - User wants different flight options
        - Budget constraints changed
        - Preferences updated after initial planning
        - Agent failures during initial execution
        
        **Process:**
        1. Resets specified agent statuses to 'pending'
        2. Clears existing recommendations for selected agents
        3. Optionally clears user selections
        4. Initiates fresh agent execution
        
        **Execution Time:** 5-10 seconds per agent (parallel execution when possible).
      operationId: rerunAgents
      tags:
        - Trips
      parameters:
        - name: tripId
          in: path
          required: true
          description: Unique trip identifier
          schema:
            type: string
            pattern: '^trip_\d+_[a-z0-9]+$'
          example: "trip_1758349943478_2fq39bbwr"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RerunAgentsRequest'
            examples:
              specific_agents:
                summary: Rerun specific agents
                value:
                  agents: ["flight", "accommodation"]
                  reason: "User wants budget airline options"
                  resetSelections: false
              all_agents:
                summary: Rerun all agents
                value:
                  agents: []
                  reason: "Complete refresh requested"
                  resetSelections: true
              failed_agent:
                summary: Retry failed agent
                value:
                  agents: ["flight"]
                  reason: "Flight agent failed due to API timeout"
                  resetSelections: false
      responses:
        '200':
          description: Agent rerun initiated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RerunAgentsResponse'
        '400':
          description: Invalid agents specified
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                success: false
                error: "Invalid agents specified"
                details: ["invalid_agent", "another_bad_agent"]
                message: "Some specified agents are not valid"
        '404':
          description: Trip not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: Execution already in progress
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                success: false
                error: "Execution in progress"
                message: "Cannot rerun agents while execution is currently in progress"
        '500':
          description: Error initiating agent rerun
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
components:
  schemas:
...
    CreateTripRequest:
      type: object
      properties:
        title:
          type: string
          maxLength: 200
          description: User-defined trip title
          example: "Romantic Paris Getaway"
        destination:
          type: string
          description: Destination city or location name
          example: "Paris"
        origin:
          type: string
          description: Origin city or location name
          example: "New York"
        departureDate:
          type: string
          format: date
          description: Trip departure date (ISO 8601 format)
          example: "2025-12-15"
        returnDate:
          type: string
          format: date
          description: Trip return date (optional for one-way trips)
          example: "2025-12-20"
        travelers:
          oneOf:
            - type: integer
              minimum: 1
              maximum: 20
              description: Simple traveler count
            - $ref: '#/components/schemas/TravelerDetails'
        preferences:
          $ref: '#/components/schemas/TripPreferences'
        collaboration:
          $ref: '#/components/schemas/CollaborationInfo'
        triggerOrchestrator:
          type: boolean
          default: true
          description: Whether to immediately start AI agent execution
      required: [destination, origin, departureDate]
...
    CreateTripResponse:
      type: object
      properties:
        success:
          type: boolean
          example: true
        data:
          type: object
          properties:
            tripId:
              type: string
              description: Unique trip identifier
              example: "trip_1758349943478_2fq39bbwr"
            status:
              type: string
              enum: [draft, planning, recommendations_ready, user_selecting, finalized, cancelled]
              description: Current trip status
            title:
              type: string
              description: Trip title
            destination:
              type: string
              description: Destination name
            origin:
              type: string
              description: Origin name
            dates:
              type: object
              properties:
                departureDate:
                  type: string
                  format: date-time
                  description: Trip departure date
                returnDate:
                  type: string
                  format: date-time
                  description: Trip return date
                duration:
                  type: integer
                  description: Trip duration in days
            travelers:
              $ref: '#/components/schemas/TravelerDetails'
            agentExecution:
              $ref: '#/components/schemas/AgentExecutionStatus'
        message:
          type: string
          description: Success message
          example: "Trip created successfully, planning in progress"
      required: [success, data, message]
...
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token authentication (not yet implemented)

# Currently no authentication required
security: []
...